ENH: Fourier-filtering for rms_slope
